project structure refactor (applied clean architecture)
