+++ app/src/main/java/nsu/titov/myconverter/presentation/fragments/ConverterFragment.kt
@@ -1,18 +1,22 @@
-package nsu.titov.myconverter
+package nsu.titov.myconverter.presentation.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import nsu.titov.myconverter.databinding.FragmentConverterBinding
+
 
 class ConverterFragment : Fragment() {
+
+    private lateinit var binding: FragmentConverterBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_converter, container, false)
+        binding = FragmentConverterBinding.inflate(inflater, container, false)
+        return binding.root
     }
 }
