+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,11 +7,11 @@
 
     <fragment
         android:id="@+id/converterFragment"
-        android:name="nsu.titov.myconverter.ConverterFragment"
+        android:name="nsu.titov.myconverter.presentation.fragments.ConverterFragment"
         android:label="fragment_converter"
         tools:layout="@layout/fragment_converter" />
     <fragment
         android:id="@+id/currencyListFragment"
-        android:name="nsu.titov.myconverter.CurrencyListFragment"
+        android:name="nsu.titov.myconverter.presentation.fragments.CurrencyListFragment"
         android:label="CurrencyListFragment" />
 </navigation>
